Feat: <LCD屏幕显示>
使用SPI3驱动1.14寸的LCD屏幕，
可以显示
	字符
	字符串
	数字
	浮点数
	点、线、方形、圆形
	图片
可以切换横屏/竖屏

还不能显示汉字。

+++ Code/SYSTEM/delay/delay.h
@@ -2,7 +2,7 @@
 #define __DELAY_H 			   
 #include <sys.h>	  
 //初始化延迟函数:1us中断一次，最小延时时间1us
-void delay_init(u8 SYSCLK);							    
+void delay_init(u32 SYSCLK);							    
 
 //系统定时器--中断处理函数
 void TimingDelay_Decrement(void);
